ximi time

+++ util/ximi.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/bash
 
 # Utility script to compile both halves at once
+SPLIT_UF=false
 
 KB=fingerpunch/ximi
 KM=apedley
@@ -13,23 +14,31 @@ LOG=$KEYMAP_DIR/compile.log
 
 echo -n "" >$LOG
 
-for SIDE in LEFT RIGHT; do
-  # cat << EOF
-  SIDE_BIN="${KB//\//_}_${KM}_${SIDE}.uf2"
-  echo Compiling $SIDE ...
-  qmk compile -kb $KB -km $KM -c -e FP_SPLIT_$SIDE=yes 2>>$LOG || exit
-  cp $BIN $SIDE_BIN
-  rm $BIN
-  qmk flash $SIDE_BIN
 
-  if [ $SIDE = LEFT ]; then
-    printf "\n-----------------------\n\n" >>$LOG
-    read -p "Press enter to continue"
-  fi
+if [ -z $SPLIT_UF ]; then
+  for SIDE in LEFT RIGHT; do
+    # cat << EOF
+    SIDE_BIN="${KB//\//_}_${KM}_${SIDE}.uf2"
+    echo Compiling $SIDE ...
+    qmk compile -kb $KB -km $KM -c -e FP_SPLIT_$SIDE=yes 2>>$LOG || exit
+    cp $BIN $SIDE_BIN
+    rm $BIN
+    qmk flash $SIDE_BIN
+
+    if [ $SIDE = LEFT ]; then
+      printf "\n-----------------------\n\n" >>$LOG
+      read -p "Press enter to continue"
+    fi
+  done
 # EOF
+else
+  echo Compiling ....
+  qmk compile -kb $KB -km $KM -c 2>>$LOG || exit
+  qmk flash $BIN
+  read -p "Press enter to continue"
+  qmk flash $BIN
+fi
 
-
-done
 
 # clear
